+++ client/components/usersposts/modifyPost.js
@@ -64,7 +64,7 @@ function pressButton() {
         city: city
       },
       imageNames: [],
-      askingPrice: parseInt(askingPrice),
+      askingPrice: parseFloat(askingPrice),
       deliveryType: {
         shipping: shipping,
         pickup: pickUp
